Company Values Example Image Sections
Added in images and sections for the examples in company values - mostly completed some issues with sizing. Currently using a for loop to simulate pulling from DB with some size restrictions but it may be more accurate if hard coded.

Attempted to apply negative margins to pull edge of each image under previous - has not worked will refactor later as minor bug - images could also be cropped to remove borders as it does not look like they are use elsewhere in the design